src/requirements.txt
gradio
google-cloud-aiplatform
httpx
python-dotenv
